Perf: cut DeepSeek-V4 SWA decode o_proj latency 1032 -> 700 us
- o_a matmul row tile 16 -> 128; the 16-row cube M starved the GEMM and
  re-streamed wo_a 32x out of L2
- tp_o_a_quant from a per-group CORE_GROUP scope to spmd(24); it ran the
  whole token loop serially on 2 of 48 AIV
- attention publish token tile 8 -> 2; 32 work items left 16 of 48
  workers idle, and splitting only the token axis keeps the vector tiles
  at their current size
- reduce-scatter window FP32 -> BF16, matching vllm-ascend, which rounds
  each rank's wo_b partial to the model dtype before the collective
  (dsa_v1.py _forward_o_proj). Halves the wire bytes; the cross-rank
  accumulation stays FP32
- split tp_o_b_dequant_publish into a 48-worker dequant and a 24-worker
  publish. Dequant is vector work that scales with cores, the put is
  fabric-bound and does not; staging the whole D row also makes each put
  one contiguous transfer instead of D/ACT_N_TILE strided fragments
- drop the padded o_local buffer and the three per-variant bridge copies;
  o_proj_reduce_scatter now writes the caller's attn_out in place

Fastest-rank median 1031.7 -> 699.6 us (a2a3, tp4, 4 dies, 128 tokens per
die, 100 rounds / 5 warmup, replayed golden).

+++ models/deepseek_v4_flash_dspark/decode_fwd.py
@@ -345,7 +345,7 @@ def decode_fwd(
     sampled_ids: pl.Out[pl.Tensor[[MAX_LOGIT_ROWS, SAMPLED_IDS_PAD], pl.INT32]],
     attention_window: pld.DistributedTensor[[ATTENTION_WINDOW_ROWS, O_GROUP_IN], pl.BF16],
     attention_signal: pld.DistributedTensor[[TP_SIZE, 1], pl.INT32],
-    o_window: pld.DistributedTensor[[O_WINDOW_ROWS, D], pl.FP32],
+    o_window: pld.DistributedTensor[[O_WINDOW_ROWS, D], pl.BF16],
     o_signal: pld.DistributedTensor[[TP_SIZE, 1], pl.INT32],
     recv_meta: pld.DistributedTensor[[N_RANKS, N_LOCAL], pl.INT32],
     recv_x: pld.DistributedTensor[[N_LOCAL * RECV_MAX, D], pl.INT8],
@@ -1181,7 +1181,7 @@ def l3_decode_fwd(
 
     attention_window_buf = pld.alloc_window_buffer([ATTENTION_WINDOW_ROWS, O_GROUP_IN], dtype=pl.BF16)
     attention_signal_buf = pld.alloc_window_buffer([TP_SIZE, 1], dtype=pl.INT32)
-    o_window_buf = pld.alloc_window_buffer([O_WINDOW_ROWS, D], dtype=pl.FP32)
+    o_window_buf = pld.alloc_window_buffer([O_WINDOW_ROWS, D], dtype=pl.BF16)
     o_signal_buf = pld.alloc_window_buffer([TP_SIZE, 1], dtype=pl.INT32)
     recv_meta_buf = pld.alloc_window_buffer([N_RANKS, N_LOCAL], dtype=pl.INT32)
     recv_x_buf = pld.alloc_window_buffer([N_LOCAL * RECV_MAX, D], dtype=pl.INT8)
@@ -1199,7 +1199,7 @@ def l3_decode_fwd(
     for rank in pl.range(pld.world_size()):
         attention_window = pld.window(attention_window_buf, [ATTENTION_WINDOW_ROWS, O_GROUP_IN], dtype=pl.BF16)
         attention_signal = pld.window(attention_signal_buf, [TP_SIZE, 1], dtype=pl.INT32)
-        o_window = pld.window(o_window_buf, [O_WINDOW_ROWS, D], dtype=pl.FP32)
+        o_window = pld.window(o_window_buf, [O_WINDOW_ROWS, D], dtype=pl.BF16)
         o_signal = pld.window(o_signal_buf, [TP_SIZE, 1], dtype=pl.INT32)
         recv_meta = pld.window(recv_meta_buf, [N_RANKS, N_LOCAL], dtype=pl.INT32)
         recv_x = pld.window(recv_x_buf, [N_LOCAL * RECV_MAX, D], dtype=pl.INT8)
